+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - structural pattern matching
    - `FST.match()` match - `f.match(Mstmt(body=[Expr(Constant(str)), ...]))`
    - `FST.search()` search - `for m in f.search(Assign([Name], ListComp))`
-   - `FST.sub()` substitute - `f.sub(MCall('old_name', _args=M(all_args=...)), 'new_name(__FST_all_args, new_kw=3)')`
+   - `FST.sub()` substitute - `f.sub(MCall('old_name', _args=M(all_args=...)), 'new_name(__FST_all_args, new_kw=newval)')`
 - `FST.find_def()` easily find (possibly nested) function and class definitions by name in a given scope by dotted name
   - added usage of this in indexing (single-element only, not slice): `method = module['class.method']`, `del someclass['method']`, etc...
 - `FST` / `fstview` interoperability for convenience
@@ -26,6 +26,8 @@
 - allow get and put as `one=True` single-item `Dict` or `MatchMapping` instead of error, makes more sense
 - simplified `trivia` option to always specify entire state, no more merging with current state
 - `Global/Nonlocal.names` views now return a single-element `FSTView` when dereferencing a single element instead of the string value at that index
+- allow delete `Raise.exc` if `cause` exists by automatically deleting `cause` first
+- allow delete `ExceptHandler.type` if `name` exists by automatically deleting `name` first
 
 ### Fixed